Ensure we copy libgc to libgc-threaded before running Guile benchmarks

+++ guile/run-benchmark.sh
@@ -2,15 +2,25 @@
 
 git clone https://github.com/ecraven/r7rs-benchmarks
 
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
+
 benchmark()
 {
   cd r7rs-benchmarks
+
+  # This is messy, but we need to copy libgc.so, which we just compiled 
+  # before running this test, to libgc-threaded.so.1 as this is what
+  # the guile binary links to.
+  cp /usr/local/lib/libgc.so /usr/local/lib/libgc-threaded.so.1
 
   # Clear the log file after previous runs.
   rm results.Guile
 
   # Run the benchmarks 30 times.
-  ./bench -r 30 guile all
+  rtprio 0 ./bench -r 30 guile all
 
   # Log output is in results.Guile, extract runtimes.
   echo "Runtime,Benchmark,Time (s)" > "../$1"
